refactor: extract PrivacyDataExportTypes.swift from PrivacyDataExporter.swift
Extracts 4 public support types (PrivacyDataExportManifest,
PrivacyDataExportSettingsSnapshot, PrivacyDataExportDiagnosticsSnapshot,
PrivacyDataSessionStream) from PrivacyDataExporter.swift into a new
PrivacyDataExportTypes.swift.

Byte-identical move; SHA-256 75044e5b1b9c27f4712193947119368ab0f1584e6f91cca34aaeefcc6b007d58
verified against origin/main:PrivacyDataExporter.swift lines 4-82.

Reduces PrivacyDataExporter.swift from 290 to 210 lines.

Tests: 667 (unchanged).
